refactor: move backend models to a the RecordShopBackend file
