--- app/src/main/java/uk/udemy/recordshop/data/model/ArtistAndAlbums.kt
@@ -1,3 +1,0 @@
-package uk.udemy.recordshop.data.model
-
-data class ArtistAndAlbums(val id: Long, val artistName: String, val albums: List<Album>)
