--- app/src/main/java/uk/udemy/recordshop/data/model/ArtistDTO.kt
@@ -1,3 +1,0 @@
-package uk.udemy.recordshop.data.model
-
-data class ArtistDTO(val id: Long, val artistName: String)
